+++ fftw/cppbuild.sh
@@ -19,15 +19,11 @@ cd fftw-$FFTW_VERSION
 
 case $PLATFORM in
     android-arm)
-        export AR="$ANDROID_BIN-ar"
-        export RANLIB="$ANDROID_BIN-ranlib"
-        export CPP="$ANDROID_BIN-cpp"
-        export CC="$ANDROID_BIN-gcc"
-        export STRIP="$ANDROID_BIN-strip"
-        export CPPFLAGS="$ANDROID_FLAGS"
-        export CFLAGS="$ANDROID_FLAGS"
-        export LDFLAGS="-Wl,--fix-cortex-a8 -z text"
-        export LIBS="-lgcc -ldl -lz -lm -lc"
+        export AR="$ANDROID_PREFIX-ar"
+        export RANLIB="$ANDROID_PREFIX-ranlib"
+        export CC="$ANDROID_CC $ANDROID_FLAGS"
+        export STRIP="$ANDROID_PREFIX-strip"
+        export LIBS="-ldl -lm -lc"
         patch -Np1 < ../../../fftw-android.patch
         ./configure --prefix=$INSTALL_PATH --disable-fortran --enable-shared --enable-threads --with-combined-threads --host="arm-linux-androideabi" --with-sysroot="$ANDROID_ROOT"
         make -j $MAKEJ V=0
@@ -37,15 +33,11 @@ case $PLATFORM in
         make install-strip
         ;;
     android-arm64)
-        export AR="$ANDROID_BIN-ar"
-        export RANLIB="$ANDROID_BIN-ranlib"
-        export CPP="$ANDROID_BIN-cpp"
-        export CC="$ANDROID_BIN-gcc"
-        export STRIP="$ANDROID_BIN-strip"
-        export CPPFLAGS="$ANDROID_FLAGS"
-        export CFLAGS="$ANDROID_FLAGS"
-        export LDFLAGS="-z text"
-        export LIBS="-lgcc -ldl -lz -lm -lc"
+        export AR="$ANDROID_PREFIX-ar"
+        export RANLIB="$ANDROID_PREFIX-ranlib"
+        export CC="$ANDROID_CC $ANDROID_FLAGS"
+        export STRIP="$ANDROID_PREFIX-strip"
+        export LIBS="-ldl -lm -lc"
         patch -Np1 < ../../../fftw-android.patch
         ./configure --prefix=$INSTALL_PATH --disable-fortran --enable-shared --enable-threads --with-combined-threads --host="aarch64-linux-android" --with-sysroot="$ANDROID_ROOT"
         make -j $MAKEJ V=0
@@ -55,15 +47,11 @@ case $PLATFORM in
         make install-strip
         ;;
      android-x86)
-        export AR="$ANDROID_BIN-ar"
-        export RANLIB="$ANDROID_BIN-ranlib"
-        export CPP="$ANDROID_BIN-cpp"
-        export CC="$ANDROID_BIN-gcc"
-        export STRIP="$ANDROID_BIN-strip"
-        export CPPFLAGS="$ANDROID_FLAGS"
-        export CFLAGS="$ANDROID_FLAGS"
-        export LDFLAGS="-z text"
-        export LIBS="-lgcc -ldl -lz -lm -lc"
+        export AR="$ANDROID_PREFIX-ar"
+        export RANLIB="$ANDROID_PREFIX-ranlib"
+        export CC="$ANDROID_CC $ANDROID_FLAGS"
+        export STRIP="$ANDROID_PREFIX-strip"
+        export LIBS="-ldl -lm -lc"
         patch -Np1 < ../../../fftw-android.patch
         ./configure --prefix=$INSTALL_PATH --disable-fortran --enable-shared --enable-threads --with-combined-threads --enable-sse2 --host="i686-linux-android" --with-sysroot="$ANDROID_ROOT"
         make -j $MAKEJ V=0
@@ -73,15 +61,11 @@ case $PLATFORM in
         make install-strip
         ;;
      android-x86_64)
-        export AR="$ANDROID_BIN-ar"
-        export RANLIB="$ANDROID_BIN-ranlib"
-        export CPP="$ANDROID_BIN-cpp"
-        export CC="$ANDROID_BIN-gcc"
-        export STRIP="$ANDROID_BIN-strip"
-        export CPPFLAGS="$ANDROID_FLAGS"
-        export CFLAGS="$ANDROID_FLAGS"
-        export LDFLAGS="-z text"
-        export LIBS="-lgcc -ldl -lz -lm -lc"
+        export AR="$ANDROID_PREFIX-ar"
+        export RANLIB="$ANDROID_PREFIX-ranlib"
+        export CC="$ANDROID_CC $ANDROID_FLAGS"
+        export STRIP="$ANDROID_PREFIX-strip"
+        export LIBS="-ldl -lm -lc"
         patch -Np1 < ../../../fftw-android.patch
         ./configure --prefix=$INSTALL_PATH --disable-fortran --enable-shared --enable-threads --with-combined-threads --enable-sse2 --host="x86_64-linux-android" --with-sysroot="$ANDROID_ROOT"
         make -j $MAKEJ V=0